refactor: migrate database schema and queries to new services directory
