+++ app/favorites.tsx
@@ -10,9 +10,9 @@ import Alert from '~/components/Alert';
 import { Container } from '~/components/Container';
 import FoodComponent from '~/components/FoodComponent';
 import TopBar from '~/components/TopBar';
-import { toggleFavorites } from '~/db/database';
-import * as schema from '~/db/schema';
 import { useDatabase } from '~/hooks/useDatabase';
+import { toggleFavorites } from '~/services/database/database';
+import * as schema from '~/services/database/schema';
 import { useSettingsStore } from '~/store/useSettingsStore';
 import { COLORS } from '~/utils/colors';
 import { cn } from '~/utils/utils';
